Create README.md
🚀 Task 2 : Analyze social media data (e.g., Twitter) to understand public sentiment towards specific topics, products, or events. Use natural language
processing (NLP) techniques to preprocess text data, extract sentiment
scores, and visualize sentiment trends over time.

In this video, I'm going through how to do a sentiment analysis project using machine learning. This is a very powerful technique that can be used to investigate the sentiment of text data.

This video is a comprehensive guide that will walk you through the entire sentiment analysis project. We'll start by reviewing the different classification models that are available, and then we'll move on to training the models. Finally, we'll use the models to perform the analysis.

#BrainwaveMatrixSolution
#Internship
#Task2
#Presentation1
#Presentation2